+++ src/SW_VegProd.c
@@ -160,6 +160,7 @@ void SW_VPD_read(
         "CO2 Biomass Coefficient 2",
         "CO2 WUE Coefficient 1",
         "CO2 WUE Coefficient 2",
+        "Spatial reference of biomass inputs (are inputs as if 100% cover)",
         "year of vegetation inputs"
     };
 
@@ -170,7 +171,7 @@ void SW_VPD_read(
     int lineno = 0;
     int index;
     // last case line number before monthly biomass densities
-    const int line_help = 29;
+    const int line_help = 30;
     double help_veg[NVEGTYPES];
     double help_bareGround = 0.;
     double litt;
@@ -199,7 +200,7 @@ void SW_VPD_read(
                                          (char *) "Invalid record in";
 
         if (lineno <= line_help) {
-            if (lineno == 1 || lineno == 29) {
+            if (lineno == 1 || lineno == 29 || lineno == 30) {
                 x = sscanf(inbuf, "%19s", vegStrs[0]);
                 expectedNumInVals = 1;
 
@@ -440,8 +441,18 @@ void SW_VPD_read(
                 }
                 break;
 
-            /* Calendar year corresponding to vegetation inputs */
+            /* Spatial reference of biomass inputs */
             case 29:
+                SW_VegProd->isBiomAsIf100Cover =
+                    sw_strtoi(vegStrs[0], MyFileName, LogInfo) ? swTRUE :
+                                                                 swFALSE;
+                if (LogInfo->stopRun) {
+                    goto closeFile;
+                }
+                break;
+
+            /* Calendar year corresponding to vegetation inputs */
+            case 30:
                 SW_VegProd->vegYear =
                     sw_strtoi(vegStrs[0], MyFileName, LogInfo);
                 if (LogInfo->stopRun) {
@@ -817,6 +828,7 @@ void SW_VPD_new_year(SW_VEGPROD *SW_VegProd, SW_MODEL *SW_Model) {
     TimeInt doy; /* base1 */
     TimeInt simyear = SW_Model->simyear;
     int k;
+    int mon;
 
     // Interpolation is to be in base1 in `interpolate_monthlyValues()`
     Bool interpAsBase1 = swTRUE;
@@ -824,10 +836,29 @@ void SW_VPD_new_year(SW_VEGPROD *SW_VegProd, SW_MODEL *SW_Model) {
     /* Monthly biomass after CO2 effects */
     double biomass_after_CO2[MAX_MONTHS];
 
+    /* Monthly biomass at 100% cover */
+    double biomassAsIf100Cover[MAX_MONTHS];
+    double litterAsIf100Cover[MAX_MONTHS];
+
 
     // Grab the real year so we can access CO2 data
     ForEachVegType(k) {
         if (GT(SW_VegProd->veg[k].cov.fCover, 0.)) {
+
+            /* Scale biomass to as if 100% cover unless provided as inputs */
+            for (mon = 0; mon < MAX_MONTHS; mon++) {
+                biomassAsIf100Cover[mon] =
+                    SW_VegProd->isBiomAsIf100Cover ?
+                        SW_VegProd->veg[k].biomass[mon] :
+                        (SW_VegProd->veg[k].biomass[mon] /
+                         SW_VegProd->veg[k].cov.fCover);
+
+                litterAsIf100Cover[mon] = SW_VegProd->isBiomAsIf100Cover ?
+                                              SW_VegProd->veg[k].litter[mon] :
+                                              (SW_VegProd->veg[k].litter[mon] /
+                                               SW_VegProd->veg[k].cov.fCover);
+            }
+
             if (k == SW_TREES) {
                 // CO2 effects on tree biomass restricted to percent live
                 // biomass, i.e., total tree biomass is constant while live
@@ -846,7 +877,7 @@ void SW_VPD_new_year(SW_VEGPROD *SW_VegProd, SW_MODEL *SW_Model) {
                     SW_VegProd->veg[k].pct_live_daily
                 );
                 interpolate_monthlyValues(
-                    SW_VegProd->veg[k].biomass,
+                    biomassAsIf100Cover,
                     interpAsBase1,
                     SW_Model->cum_monthdays,
                     SW_Model->days_in_month,
@@ -858,7 +889,7 @@ void SW_VPD_new_year(SW_VEGPROD *SW_VegProd, SW_MODEL *SW_Model) {
                 // total and live biomass are increasing
                 apply_biomassCO2effect(
                     biomass_after_CO2,
-                    SW_VegProd->veg[k].biomass,
+                    biomassAsIf100Cover,
                     SW_VegProd->veg[k].co2_multipliers[BIO_INDEX][simyear]
                 );
 
@@ -880,7 +911,7 @@ void SW_VPD_new_year(SW_VEGPROD *SW_VegProd, SW_MODEL *SW_Model) {
 
             // Interpolation of remaining variables from monthly to daily values
             interpolate_monthlyValues(
-                SW_VegProd->veg[k].litter,
+                litterAsIf100Cover,
                 interpAsBase1,
                 SW_Model->cum_monthdays,
                 SW_Model->days_in_month,